+++ plugins/discourse-events/app/models/discourse_events/events/event.rb
@@ -59,6 +59,7 @@ module DiscourseEvents
       validate :raw_invitees_length
       validate :ends_before_start
       validate :allowed_custom_fields
+      validate :url_is_a_uri, if: :url_changed?
 
       def self.attributes_protected_by_default
         super - %w[id]
@@ -256,6 +257,12 @@ module DiscourseEvents
             I18n.t("discourse_post_event.errors.models.event.ends_at_before_starts_at"),
           )
         end
+      end
+
+      def url_is_a_uri
+        return if Parser.valid_url?(url)
+
+        errors.add(:base, I18n.t("discourse_post_event.errors.models.event.invalid_url"))
       end
 
       def allowed_custom_fields
