+++ plugins/discourse-events/lib/discourse_events/events/validator.rb
@@ -137,6 +137,10 @@ module DiscourseEvents
           end
         end
 
+        if !Parser.valid_url?(extracted_event[:url])
+          @post.errors.add(:base, I18n.t("discourse_post_event.errors.models.event.invalid_url"))
+        end
+
         true
       end
